Rename rtlib as core
