+++ core/include/rt/DeformableRegistration.hpp
@@ -101,4 +101,4 @@ private:
 };
 
 }  // namespace graph
-}
+}  // namespace rt
