+++ core/include/rt/DisegniSegmenter.hpp
