+++ core/include/rt/ImageTransformResampler.hpp
@@ -46,4 +46,4 @@ private:
 };
 
 }  // namespace graph
-}
+}  // namespace rt
