+++ core/include/rt/LandmarkDetector.hpp
@@ -105,4 +105,4 @@ private:
 };
 
 }  // namespace graph
-}
+}  // namespace rt
