+++ core/include/rt/io/OBJWriter.hpp
@@ -119,4 +119,4 @@ private:
 };
 
 }  // namespace io
-}
+}  // namespace rt
