+++ core/include/rt/io/TIFFIO.hpp
