+++ core/src/CompositeTransform.cpp
