+++ core/src/DisegniSegmenter.cpp
