+++ core/src/ImageConversion.cpp
