+++ core/src/ImageTransformResampler.cpp
