+++ core/src/LandmarkWriter.cpp
