+++ core/src/OBJWriter.cpp
