+++ core/src/TIFFIO.cpp
