+++ core/src/Version.cpp.in
