+++ tests/CMakeLists.txt
@@ -21,10 +21,10 @@ set(tests
 
 foreach(src ${tests})
     get_filename_component(filename ${src} NAME_WE)
-    set(testname rtlib_${filename})
+    set(testname core_${filename})
     add_executable(${testname} ${src})
     target_link_libraries(${testname}
-        rt::rtlib
+        rt::core
         gtest_main
         gmock_main
     )
